rename to Const Affine
