+++ init.lua
@@ -9,5 +9,5 @@ require 'inn.MeanSubtraction'
 require 'inn.SpatialPyramidPooling'
 require 'inn.SpatialSameResponseNormalization'
 require 'inn.ROIPooling'
-require 'inn.SpatialConstAffine'
+require 'inn.ConstAffine'
 return inn
